Создали классы кота, собаки унаследованные от родительского класса Животные. Созданы методы и счетчики.
