+++ src/main/java/gb/Khelmyanov_HW/Lesson6/Animal.java
@@ -2,9 +2,9 @@ package gb.Khelmyanov_HW.Lesson6;
 
 public class Animal {
 
-    String name;
-    int distance;
-    int counter;
+    protected String name;
+    private int distance;
+    private int counter;
 
     public int getCounter() {
         return counter;
@@ -14,24 +14,34 @@ public class Animal {
 
     }
 
-    public Animal(String name, int distance){
-
+    public Animal(String name) {
         this.name = name;
-        this.distance = distance;
 
-        counter ++;
+        counter++;
     }
 
-    public void animalInfo(){
-        System.out.println("Животное с именем " + name);
-    }
-
-    public void animalRun(int distance) {
+    public void animalRun(int distance, Animal animal) {
         System.out.println("Животное пробежало " + distance + "м");
     }
 
-    public void animalSwim(int distance) {
+    public void animalSwim(int distance, Animal animal) {
         System.out.println("Животное проплыло " + distance + "м");
     }
 
+    public static void main(String[] args) {
+
+        Cat cat01 = new Cat("Барсик");
+        Cat cat02 = new Cat("Персик");
+        Cat cat03 = new Cat("Чертик");
+
+        Dog dog01 = new Dog("Бобик");
+        Dog dog02 = new Dog("Тузик");
+
+        cat01.animalRun(200);
+        cat02.animalSwim(100);
+
+        dog01.animalRun(499);
+        dog01.animalSwim(300);
+
+    }
 }
